Upgrade roles command - handle managed policies
1. Check that account roles have managed policies attached.
2. Check that operator roles have managed policies attached.
2. Check for missing operator roles by OCP version.

Related: SDA-7712
